UTIL - all | with "shadow-json out=yes" full PAN-OS XML config is ava…

+++ utils/lib/UTIL.php
@@ -1471,15 +1471,23 @@ class UTIL
         {
             if( $this->configOutput != '/dev/null' )
             {
-                // destroy destination file if it exists
-                if( file_exists($this->configOutput) && is_file($this->configOutput) )
-                    unlink($this->configOutput);
+                if( !PH::$shadow_json )
+                {
+                    // destroy destination file if it exists
+                    if( file_exists($this->configOutput) && is_file($this->configOutput) )
+                        unlink($this->configOutput);
 
-                //          save_to_file($fileName, $printMessage=true, $lineReturn = true, $indentingXml = 0, $indentingXmlIncreament = 1 )
-                $this->pan->save_to_file($this->configOutput, $printMessage, $lineReturn, $indentingXml, $indentingXmlIncreament);
+                    $this->pan->save_to_file($this->configOutput, $printMessage, $lineReturn, $indentingXml, $indentingXmlIncreament);
+                }
+                else
+                {
+                    //store it JSON out
+                    PH::$JSON_TMP['xmldoc'] = &DH::dom_to_xml($this->pan->xmlroot, $indentingXml, $lineReturn, -1, $indentingXmlIncreament);
+                    PH::print_stdout(PH::$JSON_TMP, false, "out");
+                    PH::$JSON_TMP = array();
+                }
             }
         }
-
 
         if( $additional_output )
         {
